+++ clients/dotnet/Boxops.Fjord.Indexer/GleanFacts.cs
@@ -1,9 +1,9 @@
 using System.Text.Encodings.Web;
 using System.Text.Json;
 
-using Fjord.Client;
+using Boxops.Fjord.Client;
 
-namespace Fjord.Indexer;
+namespace Boxops.Fjord.Indexer;
 
 /// <summary>
 /// The same facts, in Glean's JSON batch format.